requirements.txt
aiogram
python-dotenv
google-generativeai
gTTS
aiosqlite
Pillow
aiohttp-socks